[추가] HomeBodyView에 viewModel을 도입합니다.
